refactor: extract Runner class from Codecept for all Mocha interactions
- Create lib/runner.js with Runner class that owns all Mocha interactions:
  getSuites(), run(), runSuite(), runTest() backed by a single _execute() core
- Codecept.run/runSuite/runTest/getSuites now delegate to this.runner
- Remove duplicated executeTest() that was copy of run()
- Refactor check.js to use codecept.getSuites() instead of manual Mocha
- Export Runner from lib/index.js
