+++ lib/codecept.js
@@ -11,17 +11,13 @@ const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
 
 import Helper from '@codeceptjs/helper'
-import MochaFactory from './mocha/factory.js'
+import Runner from './runner.js'
 import container from './container.js'
 import Config from './config.js'
-import event from './event.js'
 import runHook from './hooks.js'
 import ActorFactory from './actor.js'
-import output from './output.js'
 import { emptyFolder } from './utils.js'
 import { initCodeceptGlobals } from './globals.js'
-import { validateTypeScriptSetup, getTSNodeESMWarning } from './utils/loaderCheck.js'
-import recorder from './recorder.js'
 import store from './store.js'
 
 import storeListener from './listener/store.js'
@@ -104,6 +100,7 @@ class Codecept {
     await this.requireModules(this.requiringModules)
     // initializing listeners
     await container.create(this.config, this.opts)
+    this.runner = new Runner(this)
     await this.runHooks()
   }
 
@@ -266,87 +263,29 @@ class Codecept {
    * @returns {Array<{title: string, file: string, tags: string[], tests: Array<{title: string, uid: string, tags: string[], fullTitle: string}>}>}
    */
   getSuites(pattern) {
-    if (this.testFiles.length === 0) {
-      this.loadTests(pattern)
-    }
-
-    const tempMocha = MochaFactory.create(this.config.mocha || {}, this.opts || {})
-    tempMocha.files = this.testFiles
-    tempMocha.loadFiles()
-
-    const suites = []
-    for (const suite of tempMocha.suite.suites) {
-      suites.push({
-        ...suite.simplify(),
-        file: suite.file || '',
-        tests: suite.tests.map(test => ({
-          ...test.simplify(),
-          fullTitle: test.fullTitle(),
-        })),
-      })
-    }
-
-    tempMocha.unloadFiles()
-    return suites
+    return this.runner.getSuites(pattern)
   }
 
   /**
-   * Execute all tests in a suite.
+   * Run all tests in a suite.
    * Must be called after init() and bootstrap().
    *
    * @param {{file: string}} suite - suite object returned by getSuites()
    * @returns {Promise<void>}
    */
-  async executeSuite(suite) {
-    return this.run(suite.file)
+  async runSuite(suite) {
+    return this.runner.runSuite(suite)
   }
 
   /**
-   * Execute a single test by its fullTitle.
+   * Run a single test by its fullTitle.
    * Must be called after init() and bootstrap().
    *
    * @param {{fullTitle: string}} test - test object returned by getSuites()
    * @returns {Promise<void>}
    */
-  async executeTest(test) {
-    await container.started()
-
-    const tsValidation = validateTypeScriptSetup(this.testFiles, this.requiringModules || [])
-    if (tsValidation.hasError) {
-      output.error(tsValidation.message)
-      process.exit(1)
-    }
-
-    try {
-      const { loadTranslations } = await import('./mocha/gherkin.js')
-      await loadTranslations()
-    } catch (e) {
-      // Ignore if gherkin module not available
-    }
-
-    return new Promise((resolve, reject) => {
-      const mocha = container.mocha()
-      mocha.files = this.testFiles
-      mocha.grep(test.fullTitle)
-
-      const done = async (failures) => {
-        event.emit(event.all.result, container.result())
-        event.emit(event.all.after, this)
-        await recorder.promise()
-        if (failures) {
-          process.exitCode = 1
-        }
-        resolve()
-      }
-
-      try {
-        event.emit(event.all.before, this)
-        mocha.run(async (failures) => await done(failures))
-      } catch (e) {
-        output.error(e.stack)
-        reject(e)
-      }
-    })
+  async runTest(test) {
+    return this.runner.runTest(test)
   }
 
   /**
@@ -356,64 +295,7 @@ class Codecept {
    * @returns {Promise<void>}
    */
   async run(test) {
-    await container.started()
-
-    // Check TypeScript loader configuration before running tests
-    const tsValidation = validateTypeScriptSetup(this.testFiles, this.requiringModules || [])
-    if (tsValidation.hasError) {
-      output.error(tsValidation.message)
-      process.exit(1)
-    }
-
-    // Show warning if ts-node/esm is being used
-    const tsWarning = getTSNodeESMWarning(this.requiringModules || [])
-    if (tsWarning) {
-      output.print(output.colors.yellow(tsWarning))
-    }
-
-    // Ensure translations are loaded for Gherkin features
-    try {
-      const { loadTranslations } = await import('./mocha/gherkin.js')
-      await loadTranslations()
-    } catch (e) {
-      // Ignore if gherkin module not available
-    }
-
-    return new Promise((resolve, reject) => {
-      const mocha = container.mocha()
-      mocha.files = this.testFiles
-
-      if (test) {
-        if (!fsPath.isAbsolute(test)) {
-          test = fsPath.join(store.codeceptDir, test)
-        }
-        const testBasename = fsPath.basename(test, '.js')
-        const testFeatureBasename = fsPath.basename(test, '.feature')
-        mocha.files = mocha.files.filter(t => {
-          return fsPath.basename(t, '.js') === testBasename || fsPath.basename(t, '.feature') === testFeatureBasename || t === test
-        })
-      }
-
-      const done = async (failures) => {
-        event.emit(event.all.result, container.result())
-        event.emit(event.all.after, this)
-        // Wait for any recorder tasks added by event.all.after handlers
-        await recorder.promise()
-        // Set exit code based on test failures
-        if (failures) {
-          process.exitCode = 1
-        }
-        resolve()
-      }
-
-      try {
-        event.emit(event.all.before, this)
-        mocha.run(async (failures) => await done(failures))
-      } catch (e) {
-        output.error(e.stack)
-        reject(e)
-      }
-    })
+    return this.runner.run(test)
   }
 
   /**
