+++ lib/command/check.js
@@ -70,15 +70,10 @@ export default async function (options) {
   if (codecept) {
     try {
       codecept.loadTests()
-      const files = codecept.testFiles
-      const mocha = Container.mocha()
-      mocha.files = files
-      mocha.loadFiles()
+      const suites = codecept.getSuites()
 
-      for (const suite of mocha.suite.suites) {
-        if (suite && suite.tests) {
-          numTests += suite.tests.length
-        }
+      for (const suite of suites) {
+        numTests += suite.tests.length
       }
 
       if (numTests > 0) {
