+++ lib/index.js
@@ -24,6 +24,7 @@ import ai from './ai.js'
 import Workers from './workers.js'
 import Secret, { secret } from './secret.js'
 import Result from './result.js'
+import Runner from './runner.js'
 
 export default {
   /** @type {typeof CodeceptJS.Codecept} */
@@ -71,7 +72,9 @@ export default {
 
   /** @type {typeof Result} */
   Result,
+
+  Runner,
 }
 
 // Named exports for ESM compatibility
-export { codecept, output, container, event, recorder, config, actor, helper, pause, within, dataTable, dataTableArgument, store, locator, heal, ai, Workers, Secret, secret, Result }
+export { codecept, output, container, event, recorder, config, actor, helper, pause, within, dataTable, dataTableArgument, store, locator, heal, ai, Workers, Secret, secret, Result, Runner }
